+++ src/renderer/components/configure-recording-set-page/index.ts
@@ -1,0 +1,1 @@
+export { ConfigureRecordingSetPage } from './configure-recording-set-page';
